# Add Parse/TryParse methods to ResourceIdentifier