--- a/sdk/core/Azure.Core/src/ResourceIdentifier.cs
+++ b/sdk/core/Azure.Core/src/ResourceIdentifier.cs
@@ -113,8 +113,6 @@ namespace Azure.Core
                     _resourceGroupName = resourceName;
                     break;
                 case SpecialType.Subscription:
-                    if (!Guid.TryParse(resourceName, out _))
-                        throw new FormatException($"The GUID for subscription is invalid {resourceName}.");
                     _subscriptionId = resourceName;
                     break;
                 case SpecialType.Provider:
@@ -138,13 +136,13 @@ namespace Azure.Core
             _initialized = true;
         }
 
-        private void Parse()
+        private string? Parse()
         {
             ReadOnlySpan<char> remaining = _stringValue.AsSpan();
 
             if (!remaining.StartsWith(SubscriptionStart.AsSpan(), StringComparison.OrdinalIgnoreCase) &&
                 !remaining.StartsWith(ProviderStart.AsSpan(), StringComparison.OrdinalIgnoreCase))
-                throw new FormatException($"The ResourceIdentifier must start with {SubscriptionStart} or {ProviderStart}.");
+                return $"The ResourceIdentifier must start with {SubscriptionStart} or {ProviderStart}.";
 
             //trim trailing '/' off the end if it exists
             remaining = remaining[remaining.Length - 1] == Separator ? remaining.Slice(1, remaining.Length - 2) : remaining.Slice(1);
@@ -152,23 +150,46 @@ namespace Azure.Core
             ReadOnlySpan<char> nextWord = PopNextWord(ref remaining);
 
             //we know we at least have 1 ResourceIdentifier in the tree here
-            ResourceIdentifierParts nextParts = GetNextParts(Root, ref remaining, ref nextWord);
+            var errorMessage = GetNextParts(Root, ref remaining, ref nextWord, out ResourceIdentifierParts? result);
+            if (errorMessage is not null)
+                return errorMessage;
+            ResourceIdentifierParts nextParts = result!.Value;
+
             while (!nextWord.IsEmpty)
             {
                 //continue to get the next ResourceIdentifier in the tree until we reach the end which will be 'this'
                 ResourceIdentifier newParent = new ResourceIdentifier(nextParts.Parent, nextParts.ResourceType, nextParts.ResourceName, nextParts.IsProviderResource, nextParts.SpecialType);
-                nextParts = GetNextParts(newParent, ref remaining, ref nextWord);
+                if (nextParts.SpecialType == SpecialType.Subscription)
+                {
+                    errorMessage = newParent.CheckSubscriptionFormat();
+                    if (errorMessage is not null)
+                        return errorMessage;
+                }
+                errorMessage = GetNextParts(newParent, ref remaining, ref nextWord, out result);
+                if (errorMessage is not null)
+                    return errorMessage;
+                nextParts = result!.Value;
             }
 
             //initialize ourselves last
             Init(nextParts.Parent, nextParts.ResourceType, nextParts.ResourceName, nextParts.IsProviderResource, nextParts.SpecialType);
+            return nextParts.SpecialType == SpecialType.Subscription ? CheckSubscriptionFormat() : null;
+        }
+
+        private string? CheckSubscriptionFormat()
+        {
+            if (_subscriptionId is not null && !Guid.TryParse(_subscriptionId, out _))
+                return $"The GUID for subscription is invalid {_subscriptionId}.";
+            return null;
         }
 
         private T GetValue<T>(ref T value)
         {
             if (!_initialized)
             {
-                Parse();
+                var error = Parse();
+                if (error is not null)
+                    throw new FormatException(error);
             }
 
             return value;
@@ -194,8 +215,9 @@ namespace Azure.Core
             return parent.ResourceType.AppendChild(resourceTypeName.ToString());
         }
 
-        private static ResourceIdentifierParts GetNextParts(ResourceIdentifier parent, ref ReadOnlySpan<char> remaining, ref ReadOnlySpan<char> nextWord)
+        private static string? GetNextParts(ResourceIdentifier parent, ref ReadOnlySpan<char> remaining, ref ReadOnlySpan<char> nextWord, out ResourceIdentifierParts? parts)
         {
+            parts = null;
             ReadOnlySpan<char> firstWord = nextWord;
             ReadOnlySpan<char> secondWord = PopNextWord(ref remaining);
 
@@ -203,16 +225,17 @@ namespace Azure.Core
             {
                 //subscriptions and resourceGroups aren't valid ids without their name
                 if (firstWord.Equals(SubscriptionsKey.AsSpan(), StringComparison.OrdinalIgnoreCase) || firstWord.Equals(ResourceGroupKey.AsSpan(), StringComparison.OrdinalIgnoreCase))
-                    throw new FormatException($"The ResourceIdentifier is missing the key for {firstWord.ToString()}.");
+                    return $"The ResourceIdentifier is missing the key for {firstWord.ToString()}.";
 
                 //resourceGroup must contain either child or provider resource type
                 if (parent.ResourceType == ResourceType.ResourceGroup)
-                    throw new FormatException($"Expected {ProvidersKey} path segment after {ResourceGroupKey}.");
+                    return $"Expected {ProvidersKey} path segment after {ResourceGroupKey}.";
 
                 nextWord = secondWord;
                 SpecialType specialType = firstWord.Equals(LocationsKey.AsSpan(), StringComparison.OrdinalIgnoreCase) ? SpecialType.Location : SpecialType.None;
-                var resourceType =parent.ResourceType.AppendChild(firstWord.ToString());
-                return new ResourceIdentifierParts(parent, new ResourceType(resourceType), string.Empty, false, specialType);
+                var resourceType = parent.ResourceType.AppendChild(firstWord.ToString());
+                parts = new ResourceIdentifierParts(parent, new ResourceType(resourceType), string.Empty, false, specialType);
+                return null;
             }
 
             ReadOnlySpan<char> thirdWord = PopNextWord(ref remaining);
@@ -224,10 +247,11 @@ namespace Azure.Core
                     if (parent.ResourceType == ResourceType.Subscription || parent.ResourceType == ResourceType.Tenant)
                     {
                         nextWord = thirdWord;
-                        return new ResourceIdentifierParts(parent, ResourceType.Provider, secondWord.ToString(), true, SpecialType.Provider);
+                        parts = new ResourceIdentifierParts(parent, ResourceType.Provider, secondWord.ToString(), true, SpecialType.Provider);
+                        return null;
                     }
 
-                    throw new FormatException($"Provider resource can only come after the root or {SubscriptionsKey}.");
+                    return $"Provider resource can only come after the root or {SubscriptionsKey}.";
                 }
 
                 ReadOnlySpan<char> fourthWord = PopNextWord(ref remaining);
@@ -235,16 +259,18 @@ namespace Azure.Core
                 {
                     nextWord = PopNextWord(ref remaining);
                     SpecialType specialType = thirdWord.Equals(LocationsKey.AsSpan(), StringComparison.OrdinalIgnoreCase) ? SpecialType.Location : SpecialType.None;
-                    return new ResourceIdentifierParts(parent, new ResourceType(secondWord.ToString(), thirdWord.ToString()), fourthWord.ToString(), true, specialType);
+                    parts = new ResourceIdentifierParts(parent, new ResourceType(secondWord.ToString(), thirdWord.ToString()), fourthWord.ToString(), true, specialType);
+                    return null;
                 }
             }
             else
             {
                 nextWord = thirdWord;
-                return new ResourceIdentifierParts(parent, ChooseResourceType(firstWord, parent, out SpecialType specialType), secondWord.ToString(), false, specialType);
+                parts = new ResourceIdentifierParts(parent, ChooseResourceType(firstWord, parent, out SpecialType specialType), secondWord.ToString(), false, specialType);
+                return null;
             }
 
-            throw new FormatException("Invalid resource id.");
+            return "Invalid resource id.";
         }
 
         private static ReadOnlySpan<char> PopNextWord(ref ReadOnlySpan<char> remaining)
@@ -480,7 +506,9 @@ namespace Azure.Core
         public static ResourceIdentifier Parse(string input)
         {
             var result = new ResourceIdentifier(input);
-            result.Parse();
+            var error = result.Parse();
+            if (error is not null)
+                throw new FormatException(error);
             return result;
         }
 
@@ -496,16 +524,17 @@ namespace Azure.Core
         /// <returns> True if the parse operation was successful; otherwise, false. </returns>
         public static bool TryParse(string input, out ResourceIdentifier? result)
         {
-            try
-            {
-                result = Parse(input);
-                return true;
-            }
-            catch
-            {
-                result = null;
+            result = null;
+            if (string.IsNullOrEmpty(input))
                 return false;
-            }
+
+            result = new ResourceIdentifier(input);
+            var error = result.Parse();
+            if (error is null)
+                return true;
+
+            result = null;
+            return false;
         }
 
         /// <summary>