@azure-sdk 2022-10-26
**API change check**

APIView has identified API level changes in this PR and created following API reviews.

[Azure.Core](https://apiview.dev/Assemblies/Review/574d4ef9317f4e489d107687632de3ed)