@JoshLove-msft 2022-10-27
> Updated the algorithm to be able to not need a try/catch block. It did add some time to the happy path, here is an example benchmark from before and after
> 
> Before
> 
> ```
> CreateGetName | 364.58 ns | 2.127 ns | 1.885 ns | 0.0129 |     720 B |
> ```
> 
> After
> 
> ```
> CreateGetName |   418.21 ns |  1.707 ns |  1.597 ns | 0.0129 |     720 B |
> ```
> 
> It isn't huge in total magnitude but it is 14% increase. Let me know if anyone sees what may have caused.

That is really interesting. I guess we could compute the cost of a null check by doing a benchmark of 1 null check vs 2 null checks.