@christothes 2022-10-27
> Updated the algorithm to be able to not need a try/catch block. It did add some time to the happy path, here is an example benchmark from before and after
> 
> Before
> 
> ```
> CreateGetName | 364.58 ns | 2.127 ns | 1.885 ns | 0.0129 |     720 B |
> ```
> 
> After
> 
> ```
> CreateGetName |   418.21 ns |  1.707 ns |  1.597 ns | 0.0129 |     720 B |
> ```
> 
> It isn't huge in total magnitude but it is 14% increase. Let me know if anyone sees what may have caused.

Should I read into the fact that the numbers are separated in different tables that these were taken from two different runs? If so, I'm not sure if they are as comparable as if run in the same benchmark session (both code paths co-exist and can be run in the same session).